Bug 1428000 - Migrate: only annotate affected files (#34)
The speed-up is massive: running the bug 1411012 migration on the Italian repo used to take ~25 seconds on my machine. After this patch it's down to 0.7 seconds. Running the migration on all locales (without pull/push) now takes less than 2 minutes, compared to 30 minutes before the patch.

In `hg annotate` data, `path` is the path relative to CWD and `abspath` is the path relative to the root of the repo. With CWD==root, they're the same and we can remove `abspath` to reduce the number of different types of paths used in the code.

committer: GitHub <noreply@github.com>