Require compare-locales in fluent.migrate (#47)
Make compare-locales a hard requirement for running and testing fluent.migrate
which already largely depends on it anyways. This reduces the complexity of
imports as well as the burden of deciding if a TestCase depends on
compare-locales or not. It also makes it less probable to accidentally skip
migration tests with `python -m unittest discover`.

The following commands may be used to run tests:

  - python setup.py test - Only syntax tests in tests/syntax will be run.
    compare-locales is not required.

  - python -m unittest discover - All tests will be run. If compare-locales is
    missing, tests will error.

    It's possible to run only syntax or only migrate tests by specifying the
    starting director on the CLI:

        python -m unittest discover -s tests/syntax
        python -m unittest discover -s tests/migrate

  - tox - Tox will create different envs one if which is py27-cl. It contains
    compare-locales and is configured to run both tests/syntax and
    tests/migrate tests. All other tests only run tests/syntax.
committer: GitHub <noreply@github.com>