bug 1438570, roll up known translation and apply transform when last dependency trickles in
If the dependencies of a transform come in different changesets,
we need to wait until they're all there.
Only if we don't find them all throughout all our changesets, we
shouldn't transform at all.
We can't start with intermediate transforms, 'cause that'd require
that we can do partial transforms, which is still bug 1321271.

Instead of just aggregating changeset.changes, keep track of
all our existing legacy strings in an extra set. That way,
we only transform files and messages touched by this very string.
That's a performance optimization. If we just went for it and
aggregated the changes and passed them to serialize_changeset,
we'd transform each file we touched before, found the snapshot
didn't change and ignore it. Keeping both datasets separate just
makes this quicker.