bug 1415844, use compare-locales for plurals order, r=stas
When transforming gecko plurals to fluent plurals,
the CLDR plurals don't matter much, but the ordering and mapping
we use in gecko does.
All those hysterical raisins are built into the plurals.py in
compare-locales, so we solely rely on that.
When serializing back, though, we use the CLDR order.

In the case where the translation doesn't have enough variants
to fill in the default form in CLDR, let's take the last variant
and copy it over. And then make that the correct default.