bug 1491738, create an empty {""} if a pattern ends or starts with \n, r=stas
Summary:
This is in line with our changes to Fluent to make inline
indents just be spaces.
It also changes the leading/trailing white-space regexes to be
strictly about start to end of the string, which is good for
multi-line transforms.
The trickery around the text vs block_text regexes is mostly
to make the \n be part of text, but it needs to be an
alternative to whitespace.

Reviewers: stas

Subscribers: flod

Tags: #secure-revision

Bug #: 1491738

Differential Revision: https://phabricator.services.mozilla.com/D11025