Bug 1596831, allow for multi-locale repos as migration targets, r=stas
This is quite a bit of impact on the validator, which assumed that
- paths are static assignments, but we need to allow for function calls
- paths for target and reference are always the same

I pondered making the check for path equality optional, but
decided to just drop it. There are other parts in the testing flow
that catch this, if required. Like running the migration with --dry-run.

Also use `locale` instead of `lang(uage)` now that we're talking
locale-dependent paths, and do so consistently.
The `--lang` commandline argument is kept for convenience, but is
considered deprecated.

Differential Revision: https://phabricator.services.mozilla.com/D53244