Bug 1615209, check for .localized when reading legacy files, r=stas
For bilingual files, entities exist regardless if they're
localized or not. That's what .localized on Entity represents.

Make sure we use that when reading legacy files, so that we
only migrate content that's actually translated.

And yes, message keys in our gettext implementation are weird,
I haven't yet figured out a better way to support contexts.

Differential Revision: https://phabricator.services.mozilla.com/D62711