Bug 1453557, foundation for transforming from Fluent sources, r=stas
This only has the bare core in `transforms`, with examples on
how to use it in `test_transform_pattern`.

We should collect some actual examples in practice before
declaring some of them to be an actual API. In the meantime,
recipes can do what the tests do, and just implement the
NodeTransformer bits to do their magic.

Differential Revision: https://phabricator.services.mozilla.com/D62741