Bug 1695751 - Make REPLACE default to normalize_printf for .properties files;r=flod
The fluent migration documentation states that to avoid issues, the caller
should always specify normalize_printf when working with .properties files.
This changes the code to default normalize_printf to true in this case. The
caller can still override the default behaviour. This might be useful for
non-Mozilla users of fluent.migrate.

Differential Revision: https://phabricator.services.mozilla.com/D106922